civicaseActivityControls - Extract "My Activities"/"Delegated". Fix binding.
== Description ==

The buttons for "My Activities" and "Delegated"

== Before ==

1. The buttons are bound to `$scope.involving.{foo}`.  When composing the
request, the `$scope.involving.{foo}` is read and mapped to API params.

2.  When a user clicks "Clear filters", the filters remain active.  This is
because `$scope.involving` is not part of `$scope.filters`.  When
`civicaseActivityFilters` tries to reset parameters, the `$scope.involving`
is not visible.

3. The user can select zero, one, or two of the filters "My Activities" and
"Delegated".  However, selecting both filters is nonsensical because the
filter criteria for "My Activities" and "Delegated (Other People's
Activities)" are separate.

4. The toggled/untoggled styling is unique to the widget.

== After ==

1. The buttons are bound to `$scope.filters['@contactInvolving']`. When
composing the request, the `$scope.filters['@contactInvolving']` is mapped
to API params.

2. When a user clicks "Clear Filters", the filters are reset.

3.  The user can select zero or one of the filters "My Activities" and
"Delegated".

4.  The toggled/untoggled styling is more consistent with Bootstrap /
ui-bootstrap defaults.