civicaseUploader - Add size. Add labels for quasi-columns.
The mockup from [here](https://www.awesomescreenshot.com/image/2673810/b40d7c3c56de8a003c91b259108e7f88)
is quasi-tabular (where the name, size, and action-button are supposed to
align).  I'm not sure of a great way to get that layout, but this provides some labels and
placeholder classes so that a themer can do it.