getActivityFeedUrl - Don't trample current URL
This function builds a new URL for use for hyperlinks.  However, it also has
the side-effect of chaning the current URL, which leads to some peculiar bugs.

(Ex: Start out at `civicrm/case/a/#/case/list` and then click on a case.
The UI shows the expected `summary` tab, but the URL shows `tab=activities`.
If you reload the page, it opens the `activities` tab.)