CATL-939: Build Case Dashboard Table Columns Dynamically
Added a new API action, 'gettablesummary' that returns the list of cases and
the columns that should appear on case table summary used to list cases.

Columns in case dashboard were hard-coded on angular template. Changed the
template so the columns are now added dynamically from the list of headers
returned by the gettablesummary API call.