CATL-939: Remove Double Cycle on Headers od Case List Table
Two cycles were being used to show headers of columns on case list table. This
had the side effect of having all unsortable columns be shown at the end of
the table, which might not be what is intended.

Fixed by validating if field is sortable oncivicaseSortheader directive to add
sorting functionality.