CATL-955: Allow to Remove Locked Contacts From a Case
Although you could remove contacts from the form, this was not translating to
the backend. Thus, when the form was opened again, the same contacts were
shown to be locked out of the case.

Fixed by deleting all contacts from the case being updated and inserting all
contacts into case lock table.