CATL-955: Fix Front-end Issues
Even though the checkbox for a locked case was not shown, selecting all cases
was allowing to do bulk operations on locked cases. Also, the lock icon was
being shown, when it shouldn't.

Fixed by skipping locked cases when select all checkbox is used and hiding the
lock for locked cases.